Modificações apontadas no pull request
