+++ src/main/java/com/ufcg/es/biblioconex/service/AlunoCriarService.java
@@ -1,4 +1,4 @@
-package com.ufcg.es.biblioconex.service.aluno;
+package com.ufcg.es.biblioconex.service;
 
 import com.ufcg.es.biblioconex.dto.AlunoPostPutRequestDTO;
 import com.ufcg.es.biblioconex.model.Aluno;
@@ -6,5 +6,4 @@ import com.ufcg.es.biblioconex.model.Aluno;
 @FunctionalInterface
 public interface AlunoCriarService {
 
-    Aluno criar(AlunoPostPutRequestDTO alunoPostPutRequestDto);
 }
